requirements.txt
pandas
yfinance
psycopg2-binary
python-dotenv
fastapi
uvicorn[standard]
pydantic
requests